Don’t cache.
Not caching makes it possible for you to change output size
and text, and not having to reload twice.
Yet changing the collage colour is still not updated.

It is updated on the server though. So apparently the swarmbot
script uses a cached representation of these images?

Also, Pango requires an integer image size

Also, I’m now using my given name in git